Refactor pub client by removing unused circuit breaker and retry logic
